colors 0

+++ Practica1/Interfaces/src/main/java/com/example/interfaces/IFont.java
@@ -1,7 +1,7 @@
 package com.example.interfaces;
 
 public interface IFont {
-    int getSize();
+    float getSize();
 
     boolean isBold();
